DB Improvements:  - Names check are made before sending queries  - Tables edited with more strict checcks  - Cleaned up some code  - Inserted new function: find"food"intables. It's use to check if a food can be deleted without creating problems. Manager imrpovement:  - Ingredients and meals are not deleted if they're used somewhere Creators:  - You cannot delete meals and ingredients who are used.

+++ src/main/java/pickyeater/UI/pages/choosers/IngredientChooser.java
@@ -102,6 +102,10 @@ public class IngredientChooser extends JDialog {
                     Ingredient selectedIngredient = searchedIngredients.get(selectedIndex);
                     FoodPopupMenu popupMenu = new FoodPopupMenu();
                     popupMenu.addDeleteListener(l -> {
+                        if(ingredientsSearcherExecutor.isIngredientUsed(selectedIngredient)){
+                            JOptionPane.showMessageDialog(parent,"Cannot delete this ingredient as it's being used!");
+                            return;
+                        }
                         int choice = JOptionPane.showConfirmDialog(parent,"Are you sure you want to delete it?");
                         if(choice != JOptionPane.YES_OPTION) return;
                         ingredientsSearcherExecutor.deleteIngredient(selectedIngredient);
