+++ src/main/java/pickyeater/UI/pages/choosers/MealsChooser.java
@@ -111,6 +111,10 @@ public class MealsChooser extends JDialog {
                     Meal selectedMeal = searchedMeals.get(selectedIndex);
                     FoodPopupMenu popupMenu = new FoodPopupMenu();
                     popupMenu.addDeleteListener(l -> {
+                        if(mealSearcherExecutor.isMealUsed(selectedMeal)){
+                            JOptionPane.showMessageDialog(parent,"Cannot delete this meal as it's being used!");
+                            return;
+                        }
                         int choice = JOptionPane.showConfirmDialog(parent,"Are you sure you want to delete it?");
                         if(choice != JOptionPane.YES_OPTION) return;
                         mealSearcherExecutor.deleteMeal(selectedMeal);
